deforestation_detection_using_deep_learning/requirements.txt
tensorflow>=2.8.0
keras>=2.8.0
numpy>=1.19.5
matplotlib>=3.3.4
pandas>=1.2.4
opencv-python>=4.5.3
scikit-learn>=0.24.2
pillow>=8.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
aiofiles>=0.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 